requirements.txt
# Dependencies
discord.py 
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Dev dependencies
black
